db/hints: Move the rebalancing logic to hint_storage
This commit continues modularizing manager.hh.
