+++ configure.py
@@ -939,6 +939,7 @@ scylla_core = (['message/messaging_service.cc',
                 'db/functions/function.cc',
                 'db/hints/internal/hint_endpoint_manager.cc',
                 'db/hints/internal/hint_sender.cc',
+                'db/hints/internal/hint_storage.cc',
                 'db/hints/manager.cc',
                 'db/hints/resource_manager.cc',
                 'db/hints/host_filter.cc',
